House Keeping and implementation
Implemented tool tips for enemy variables. Adjusted lighting so the dark is dark and light is light but subtle. ranges are tuned adjust particle system for fire to look better and more realistic from distance. added some torches for a little guidance in the dark. Tool tips are on all variables in enemy AI script now except Components.