Implemented One weapon and gunpickup
Added a new pistol and restored gunpickup functionality.  Adjusted gravity, and it is now corrected.  There was a value that wasn't (encapsulated).  I left it a bit floaty for now, because it is mars after all.  I'm working on disabling sound while jumping (properly functions) and restoring the sound when you land (WIP).