Grenade Audio / Damage Bug fixed
The explosion sphere had a destroy time of 0 seconds.  This caused very hit or miss chances to actually spawn the explosion sphere to deal damage.  Changing it to a 0.5 second destroy time (with the mesh render turned off) allows enemies to be damaged and the audio to play.  Very cool!
The grenade also now looks like a grenade.  I thought the sphere was fine *shrug*